+++ articleapp/models.py
@@ -1,0 +1,36 @@
+from django.db import models
+
+
+class Author(models.Model):
+    name = models.TextField()
+
+
+class Article(models.Model):
+    class Meta:
+        db_table = "article"
+
+    title = models.CharField(max_length=50)
+    user = models.ForeignKey(Author, on_delete=models.CASCADE)
+    category = models.TextField(null=True)
+    content = models.TextField(null=True)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    article_hits = models.IntegerField(default=0)
+    objects = models.Manager()
+
+    def __str__(self):
+        return self.title
+
+
+class ArticleHits(models.Model):
+    class Meta:
+        db_table = 'article_hits'
+        verbose_name = '조회수 모델'
+        verbose_name_plural = '조회수 모델'
+
+    client_ip = models.TextField(db_index=True)
+    date = models.DateTimeField(auto_now_add=True, verbose_name='조회날짜')
+    article = models.ForeignKey(Article, on_delete=models.CASCADE, verbose_name='게시글')
+
+    def __str__(self):
+        return str(self.article.id)
