--- articleapp/models/Author.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-
-class Author(models.Model):
-    name = models.TextField()
