+++ articleapp/tests/test_service_article.py
@@ -11,7 +11,8 @@ from articleapp.models import Article, Author
 from articleapp.services.service_article import (
     create_article, delete_article, read_all_article, read_article_by_title,
     read_article_by_user, read_article_containing_username,
-    read_category_article, update_article, read_article_within_a_specific_period)
+    read_article_within_a_specific_period, read_category_article,
+    update_article)
 
 
 class TestView(TestCase):
